backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
torch>=2.0.0
transformers>=4.35.0
groq>=0.4.0
requests>=2.31.0
huggingface-hub>=0.19.0
Pillow>=10.0.0
